cust_defconfig : enable binfmt_misc
binfmt_misc (Miscellaneous Binary Format) is a capability of the Linux kernel which allows arbitrary executable file formats to be recognized and passed to certain user space applications, such as emulators and virtual machines.
It is one of a number of binary format handlers in the kernel that are involved in preparing a user-space program to run.

# Conflicts:
#	arch/arm64/configs/cust_defconfig